Backend Integration
Now that the window is working, I connected our backend to the frontend. I imported the PlantDataManager class we made earlier and set it up so that it automatically loads our plants_dataset.csv (this file is yet to be created) file right when the app starts. I did this inside the __init__ method so the data is ready to go immediately. This is better than making the user click a 'Load Data' button every time they open the app.