Image Selection Feature
I added the first real feature for the user: the ability to actually pick a picture. I used the tkinter.filedialog module to make a 'Select Plant Image' button. When you click it, it opens a window so you can browse your computer and pick a .jpg or .png file. I also added a label next to it that updates to show the file path you picked, so you know it actually worked.