Interactive Elements & Polish
To finish up the GUI requirements, I added a few more buttons. I made a 'Reset Selection' button that clears the image you picked if you want to start over, and an 'Exit' button that safely closes the program. I also added a dummy 'Save Report' button in the results window. This makes sure we definitely hit the requirement of having at least 5 interactive elements for the milestone rubric.